+++ src/components/Roadmap/common/RoadmapInfoForm/index.tsx
@@ -1,56 +1,14 @@
-import { Controller, type SubmitHandler, useForm } from 'react-hook-form';
-import { useRouter } from 'next/router';
-import { useGetRoadmapsById, usePostRoadmaps, usePostRoadmapsById } from '@/api/hooks/roadmap';
-import type { PostRoadmapsRequest } from '@/api/roadmap/type';
+import { Controller } from 'react-hook-form';
 import * as Styled from '@/components/Roadmap/common/RoadmapInfoForm/style';
 import Button from '@/components/common/Button';
 import Flex from '@/components/common/Flex';
 import Input from '@/components/common/Input';
 import RadioButton from '@/components/common/RadioButton';
 import TextArea from '@/components/common/TextArea';
-import TILY_LINKS from '@/constants/links';
-import useQueryParam from '@/hooks/useQueryParam';
+import { useRoadmapInfo } from './useRoadmapInfo';
 
 const RoadmapInfoForm = () => {
-  const roadmapId = useQueryParam('roadmapId');
-  const router = useRouter();
-  const path = router.asPath.split('/').at(-1);
-
-  const { postRoadmapsAsync, isLoading: createLoading } = usePostRoadmaps();
-  const { postRoadmapsByIdAsync, isLoading: editLoading } = usePostRoadmapsById();
-  const { data } = useGetRoadmapsById({ roadmapId: Number(roadmapId) });
-
-  const {
-    control,
-    handleSubmit,
-    formState: { errors },
-  } = useForm({
-    defaultValues: {
-      name: data?.result.name || '',
-      description: data?.result.description || '',
-      isPublic: data?.result.isPublic || true,
-      isRecruit: data?.result.isRecruit || true,
-    },
-    mode: 'onSubmit',
-  });
-
-  const onSubmit: SubmitHandler<PostRoadmapsRequest> = async (formData) => {
-    if (path === 'create') {
-      const { isRecruit, ...restFormData } = formData;
-      const data = await postRoadmapsAsync({ body: restFormData });
-
-      if (data?.code === 200) {
-        router.push(TILY_LINKS.manageGroupInfo(data?.result.id));
-      }
-      return;
-    }
-    if (path === 'roadmapInfo') {
-      await postRoadmapsByIdAsync({ roadmapId: Number(roadmapId), body: formData });
-
-      return;
-    }
-  };
-
+  const { path, data, createLoading, editLoading, control, handleSubmit, errors, onSubmit } = useRoadmapInfo();
   return (
     <>
       <Styled.RoadmapInfoForm onSubmit={handleSubmit(onSubmit)}>
